Remove CPU-usage info out of XBMC
Too many times people complain that XBMC takes up 100% CPU-time. What they don't know is, that refreshing the CPU-info is the reason for it. That being said, the function is clearly not functional from a user point of view aka meeting the expectations from the users.

By removing it, hopes are up that those complains won't show up forever in the bugtracer