Remove python as debootstrap dependency
Due to dependency issues in debian jessie it's
impossible to install python in the debootstrap
stage. Removing lsb-release and python-tz allows
debootstrap to install without python.

The xbian-package-repo package has been updated a
while ago to work without lsb-release so we install
the latest version of it. xbian-update will eventually
install lsb-release as its own dependency.